logmsg: create a non-static inline version of some functions:
The affected functions are:
* log_msg_get_value()
* log_msg_get_value_by_name()

The old version are available with a __ prefix.

Using the static inline modifier they are not exported to other
libraries. This patch applies this modification through the whole
source code so the performance shouldn't be affected. Bazsi mentioned
some micro benchmarks which showed the necessity of the static inline
modifier in this case.